fix: fix the comparison logic, see below:
in case our system stup to terminate after 4h and the cron check is running
every 15 minutes, we want the system to be terminated in the timeframe
between 4h 0m and 4h 15m.

When we use the '<' instead of '<=' system will be terminated in the timeframe
between 5h 0m and 5h 15m.