feat: refactoring the monitor script
switched it to work with DateTime objects instead of DateTimeInterval objects.
Cause DateTime objects are easily comparable in php, also in php5.x.

Switching to DateTime object also allowed to use intervals as described here
https://www.php.net/manual/en/datetime.formats.relative.php

Script is now assuming that expiry interval will be provided as a single number
and a unit without sign and without whitespace, some examples that will work:

1hour
21minute
173minutes
25days

Some examples that will not work, even though those are correct for PHP:
1 hour
1hour1minute
-4days
+10minutes