feat: allow stdin data feed
added support to digest log data from the stdin.
This is needed cause on some systems we have access_log redirected to sys_log
and on top of that some systems have healthchecks or monitoring accessing
them all the time, means we need to pre-filter access log data with cli tools.

It is hard to build a flexible filtering mechanism in php code, therefore it
was decided to support stdin data stream so that data can be pre-filtered
before executring the script with tools better intended for that.

It is clear that current implementation leaves some space for improvement,
at the same time it was done in the scope of improvement day and time was a bit
pressing. Future work should re-thing possible duplication reduction withtin
 the php code.