fix: force the filter to stick to time
problem is that in syslog we had pid in the square brackets as well.
That makes detection of actual time in stdin quite hard.
This commit is based on a very shaky assumption that time always includes
a colon sign, i.e. th `:`.

Changind for both log analysis and stdin parser cause it should affect
both.

At this time it is not clear how we can solve it better