AdHoc (refector) Refector code so it deduplicates
There is a divergation in code/behavior between checking for a date. The
behavior might be different if the input is readed from a file or if it
is passed to the programm via standard input.

There was a very basic aproach on how to recognize a date in a
log-string. This was implemented for stdin and log file phrasing in
3683ef554c3030cbc9405bcf3e4051dcf6b41354. About a month later this
aproach was changed but only for stdin
2fe6ceba1a69ed8b90ada5472c0d2bfb19bfe95e

I have now refectored this phrasing in a seperate function so the code
is not duplicated and changed the matching behavior to the later aproach
with was only implemented for stdin.